Correct an error in the text of 91:4.3 Here we have British billions (10^12), not American (10^9) The American text says:
"Your own solar center radiates almost one hundred billion tons of actual matter annually, ..."

This is a quote from page 98 of Eddington's "Stars and Atoms", which says:

"... the heat emitted by the sun each year has a mass of 120 billion tons;..."

This corresponds to the figure obtained by a modern calculation, see e.g. here:
http://www.slate.com/blogs/bad_astronomy/2014/07/14/solar_wind_versus_fusion_how_does_the_sun_lose_mass.html

However, in the Preface to "Stars and Atoms" Eddington clearly says:

"The English billion, trillion, &c. (10^12, 10^18, &c.) are used."

So, the English billion is "million million". But throughout the Urantia Papers the word "billion" is used many times
(91 times in 74 paragraphs) and it is always in the American sense, i.e. as 10^9 (thousand million).

Therefore, all translations (e.g. both Russian UF and NYC) which translate it as corresponding to the American billion are
in error and ought to be corrected.